delete one sentence, I write it in description
"a random person may happen to get interested in cs_internship and want to enter one of the courses. Queue executive will be asked to add such a person to pre-queue. by either a mentor or the interns in the queue telegram group. " this sentence is not true because Queue executive will be asked to add such a person to pre-queue.